+++ CHANGES.md
@@ -17,6 +17,7 @@ Unreleased
 
 - Add `CamelCaseSchemaTransform.to_camelcase` method.
 - Add `py_gql.exts.scalars.JSONString` scalar type.
+- Add `py_gql.exts.scalars.DateTime`, `py_gql.exts.scalars.Date` and `py_gql.exts.scalars.Time` scalar type based on [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601).
 
 [0.6.1](https://github.com/lirsacc/py-gql/releases/tag/0.6.1) - 2020-04-01
 --------------------------------------------------------------------------